requirements.txt
fastapi==0.114.1
uvicorn[standard]==0.30.6
pydantic==2.8.2
onnxruntime
opencv-python-headless
numpy
pillow==10.4.0
python-multipart==0.0.9
albumentations
torch
ultralytics==8.2.58